==> Codes/03-cumulative-production-by-company.ipynb <==
import pandas as pd
import numpy as np
from IPython.display import display, HTML

prod = pd.read_csv('../Files/processed-data/prod_summ.csv').ix[:,['WELLID', 'CUMULATIVEOIL']]
well_header = pd.read_csv("../Files/processed-data/well_header.csv").ix[:,["WELLID","CURRENTOPERATOR"]]
well_production = pd.merge(prod,well_header)

by_company = well_production.groupby('CURRENTOPERATOR')
cummulated_oil = by_company.sum()
cummulated_oil['Percentage'] = cummulated_oil['CUMULATIVEOIL']/cummulated_oil['CUMULATIVEOIL'].sum()
cummulated_oil = cummulated_oil.sort_values(by = 'Percentage', ascending = False)
cummulated_oil['Cummulated Percentage'] = np.cumsum(cummulated_oil['Percentage'])

display(cummulated_oil[cummulated_oil['CUMULATIVEOIL'] > 10000000])
