==> Codes/09-Interactive-plots-and-heatmaps.ipynb <==
import pandas as pd 
import numpy as np
from pandas import DataFrame, Series
import plotly.plotly as py
import plotly.tools as tls
import gmaps as gmaps
import csv
from BeautifulSoup import BeautifulSoup
#path_processed = 'C:/Users/Aayush - Carlson/Documents/GitHub/msba-python-project/Files/processed-data/'

## Earthquake Yearly Interactive Plot 

eq_data = pd.read_csv('../Files/processed-data/completeEarthquakeFreq.csv')
eq_data.head()
eq_data['origintime'] = pd.to_datetime(eq_data['origintime'])
eq_data['YEAR'] =  eq_data.origintime.apply(lambda x: x.year)
eq_data = eq_data[eq_data['YEAR'] != 2016]
tls.set_credentials_file(username='aayushmnit', api_key='lft7xcme6n')
data = []
layout = dict(
    title = 'Earthquake Yearly data',
    # showlegend = False,
    autosize = False,
    width = 1000,
    height = 1200,
    hovermode = False,
    legend = dict(
        x=0.7,
        y=-0.1,
        bgcolor="rgba(255, 255, 255, 0)",
        font = dict( size=11 ),
    )
)
years = eq_data['YEAR'].unique()
counts = list(eq_data.groupby('YEAR').size())
for i in range(len(years)):
    geo_key = 'geo'+str(i+1) if i != 0 else 'geo'
    lons = list(eq_data[ eq_data['YEAR'] == years[i] ]['longitude'])
    lats = list(eq_data[ eq_data['YEAR'] == years[i] ]['latitude'])
    # Walmart store data
    data.append(
        dict(
            type = 'scattergeo',
            showlegend=False,
            lon = lons,
            lat = lats,
            geo = geo_key,
            name = years[i],
            marker = dict(
                color = "rgb(0, 0, 255)",
                opacity = 0.5
            )
        )
    )
    # Year markers
    data.append(
        dict(
            type = 'scattergeo',
            showlegend = False,
            lon = [-78],
            lat = [47],
            geo = geo_key,
            text = [str(years[i]) + ',' + str(counts[i])],
            mode = 'text',
        )
    )
    layout[geo_key] = dict(
        scope = 'usa',
        showland = True,
        landcolor = 'rgb(229, 229, 229)',
        showcountries = False,
        domain = dict( x = [], y = [] ),
        subunitcolor = "rgb(255, 255, 255)",
        
    )


def draw_sparkline( domain, lataxis, lonaxis ):
    ''' Returns a sparkline layout object for geo coordinates  '''
    return dict(
        showland = False,
        showframe = False,
        showcountries = False,
        showcoastlines = False,
        domain = domain,
        lataxis = lataxis,
        lonaxis = lonaxis,
        bgcolor = 'rgba(255,200,200,0.0)',
        scale = 10
    )

z = 0
COLS = 7
ROWS = 9
for y in reversed(range(ROWS)):
    for x in range(COLS):
        geo_key = 'geo'+str(z+1) if z != 0 else 'geo'
        layout[geo_key]['domain']['x'] = [float(x)/float(COLS), float(x+1)/float(COLS)]
        layout[geo_key]['domain']['y'] = [float(y)/float(ROWS), float(y+1)/float(ROWS)]
        z=z+1
        if z > 63:
            break

fig = { 'data':data, 'layout':layout }
py.iplot( fig, filename='Oklahoma Earthquakes', height=1200, width=1000 )

## Heatmap of Eartquake Combined

import gmaps
gmaps.configure(api_key="AIzaSyA7d5cMaiUJCd-EjY5FB1IfdhF-HWu5C4k") # Your Google API key
data_earthquake = list()

#getting data in right format
for i in range(eq_data.shape[0]):
    data_earthquake.append(((eq_data.latitude[i],eq_data.longitude[i])))

m = gmaps.Map()
m.add_layer(gmaps.Heatmap(data=data_earthquake))
m

## Heatmaps of Wells across Okhlahoma

prod = pd.read_csv('../Files/processed-data/prod_summ.csv').\
ix[:,["API","WELLID","FIRSTPRODDATE","LASTPRODDATE","CUMULATIVEOIL","FIRSTTWELVEMONTHOIL","FIRSTTWENTYFOURMONTHOIL"]]

well_header = pd.read_csv("../Files/processed-data/well_header.csv").\
ix[:,["WELLID","COUNTY","SURFACELATITUDE","SURFACELONGITUDE"]]

county = pd.read_csv("../Files/processed-data/county.csv").ix[:,["Geo_ID","State","COUNTY"]]
combined_data = pd.merge(pd.merge(prod,well_header),county)
combined_data.FIRSTPRODDATE = pd.to_datetime(combined_data.FIRSTPRODDATE)
combined_data.LASTPRODDATE = pd.to_datetime(combined_data.LASTPRODDATE)
combined_data1 = combined_data.ix[(combined_data.SURFACELONGITUDE < -93) & (combined_data.SURFACELONGITUDE > -104)].copy()
combined_data1.index = range(0,combined_data1.shape[0])

data_well = list()

#getting data in right format
for i in range(combined_data1.shape[0]):
    data_well.append(((combined_data1.SURFACELATITUDE[i],combined_data1.SURFACELONGITUDE[i])))

m = gmaps.Map()
m.add_layer(gmaps.Heatmap(data=data_well))
m

## County level Heatmap of Oil production

combined_data1 = combined_data.groupby('Geo_ID',as_index=False)['CUMULATIVEOIL'].sum()
combined_data1['FIPS'] = combined_data1.Geo_ID.apply(lambda x: x[9:14])
combined_data1['FIPS'] = combined_data1['FIPS'].astype(int)

unemployment = {}
min_value = 100; max_value = 0
for row in range(combined_data1.shape[0]):
    try:
        unemployment[str(combined_data1.FIPS[row])] = combined_data1.CUMULATIVEOIL[row]
    except:
        pass

# Load the SVG map
svg = open('../Files/processed-data/counties.svg', 'r').read()

# Load into Beautiful Soup
soup = BeautifulSoup(svg, selfClosingTags=['defs','sodipodi:namedview'])

# Find counties
paths = soup.findAll('path')

# Map colors
#colors = ["#F1EEF6", "#D4B9DA", "#C994C7", "#DF65B0", "#DD1C77", "#980043"]
colors = ['#d7191c','#fdae61','#a6d96a','#1a9641']
# County style
path_style  = 'font-size:12px;fill-rule:nonzero;stroke:#FFFFFF;stroke-opacity:1;stroke-width:0.1;stroke-miterlimit:4;stroke-dasharray:none;stroke-linecap:butt;marker-start:none;stroke-linejoin:bevel;fill:'

# Color the counties based on unemployment rate
for p in paths:     
    if p['id'] not in ["State_Lines", "separator"]:
        # pass
        try:
            rate = unemployment[p['id']]
        except:
            continue
        if rate > 57081960:
            color_class = 0
        elif rate > 41243770:
            color_class = 1
        elif rate > 32847920:
            color_class = 2
        else:
            color_class = 3
 
        color = colors[color_class]
        p['style'] = path_style + color

import io
with io.FileIO('../Plots/county_level_production_heatmap.svg', 'w') as file:
    file.write(soup.prettify())

## County level Heatmap of Earthquake

eq_grouped = eq_data.groupby('county',as_index=False)['id'].count()
eq_grouped.columns = ["COUNTY","eq_combined"]
print eq_grouped.shape
county.COUNTY = county.COUNTY.str.replace(" ","")
eq_grouped = pd.merge(eq_grouped,county,on = "COUNTY",how = "left")
eq_grouped.head()

eq_grouped = eq_grouped.groupby('Geo_ID',as_index=False)['eq_combined'].sum()
eq_grouped['FIPS'] = eq_grouped.Geo_ID.apply(lambda x: x[9:14])
eq_grouped['FIPS'] = eq_grouped['FIPS'].astype(int)

unemployment = {}
min_value = 100; max_value = 0
for row in range(eq_grouped.shape[0]):
    try:
        unemployment[str(eq_grouped.FIPS[row])] = eq_grouped.eq_combined[row]
    except:
        pass

# Load the SVG map
svg = open('../Files/processed-data/counties.svg', 'r').read()

# Load into Beautiful Soup
soup = BeautifulSoup(svg, selfClosingTags=['defs','sodipodi:namedview'])

# Find counties
paths = soup.findAll('path')

# Map colors
#colors = ["#F1EEF6", "#D4B9DA", "#C994C7", "#DF65B0", "#DD1C77", "#980043"]
colors = ['#d7191c','#fdae61','#a6d96a','#1a9641']
# County style
path_style  = 'font-size:12px;fill-rule:nonzero;stroke:#FFFFFF;stroke-opacity:1;stroke-width:0.1;stroke-miterlimit:4;stroke-dasharray:none;stroke-linecap:butt;marker-start:none;stroke-linejoin:bevel;fill:'

# Color the counties based on unemployment rate
for p in paths:     
    if p['id'] not in ["State_Lines", "separator"]:
        # pass
        try:
            rate = unemployment[p['id']]
        except:
            continue
        if rate > 170.75:
            color_class = 0
        elif rate > 28:
            color_class = 1
        elif rate > 7.75:
            color_class = 2
        else:
            color_class = 3
 
        color = colors[color_class]
        p['style'] = path_style + color

import io
with io.FileIO("../Plots/county_level_earthquake_heatmap.svg", "w") as file:
    file.write(soup.prettify())
